chore: Renames overview section stylesheet file
