+++ apps/site/components/OverviewSection/OverviewSectionItem.tsx
@@ -1,12 +1,8 @@
 import React from 'react'
-import styles from './OverviewSection.module.css'
+import styles from './overview-section.module.css'
 
-const OverviewSectionItem = ({children}) => {
-  return (
-    <article className={styles.overviewSectionItem}>
-      {children}
-    </article>
-  )
+const OverviewSectionItem = ({ children }) => {
+  return <article className={styles.overviewSectionItem}>{children}</article>
 }
 
 export default OverviewSectionItem
